% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_import.R
\encoding{UTF-8}
\name{ctrl_c}
\alias{ctrl_c}
\title{Exporter des données vers le presse-papier}
\usage{
ctrl_c(data, col.names = TRUE, row.names = FALSE, memory = 2^5, dec = ",", ...)
}
\arguments{
\item{data}{les données à exporter.}

\item{col.names}{Booléen indiquant si l'on souhaite exporter les noms des colonnes (par défaut
\code{col.names = TRUE})}

\item{row.names}{Booléen indiquant si l'on souhaite exporter les noms des lignes (par défaut
\code{row.names = FALSE})}

\item{memory}{entier indiquant la taille maximale de l'objet à copier dans le presse-papier. Par défaut
la limite est de 32 Kb (=2^5 Kb).}

\item{dec}{le caractère utilisé comme séparateur décimal dans le tableur où l'on souhaite
exporter les données. Par défaut la virgule est utilisée : utiliser le point pour si l'on utilise
tableur non francophone.}

\item{...}{autres paramètres de \code{\link{read.table}} (uniquement si les données
du presse-papie contiennent plusieurs colonnes).}
}
\description{
Fonction qui permet d'exporter des données (un vecteur ou un \code{\link{data.frame}} par exemple)
vers le presse-papier (pour ensuite les coller dans un tableur par exemple).
}
\details{
Il est possible qu'il y ait une ligne en trop si la fonction est utilisée sous Mac Os.
}
\examples{
ctrl_c(c("line 1", "line 2 \n and line 3"))
}
\seealso{
\code{\link{ctrl_v}} pour importer depuis le presse-papier.
}
