% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\encoding{UTF-8}
\name{data_to_ts}
\alias{data_to_ts}
\alias{ymd_ts}
\alias{dmy_ts}
\title{Convertisseur tableau en ts}
\usage{
ymd_ts(x, col_date = 1, sep_date = "/", frequence = NULL)

dmy_ts(x, col_date = 1, sep_date = "/", frequence = NULL)
}
\arguments{
\item{x}{table contenant au moins deux colonnes dont une avec des dates.}

\item{col_date}{numéro ou nom de la colonne contenant les dates (par défaut c'est la première colonne).}

\item{sep_date}{caractère utilisé pour séparé le jour, le mois et l'année (par défaut \code{col_date = "/"}
c'est-à-dire que les dates sont au format JJ/MM/AAAA ou AAAA/MM/JJ).}

\item{frequence}{fréquence de l'objet \code{\link[stats]{ts}}. Par défaut la fonction fait une recherche automatique
(il peut y avoir des problèmes si l'on n'a pas d'année entière connue).}
}
\value{
Un \code{\link[stats]{ts}}.
}
\description{
Fonctions qui permettent de convertir des tableaux en \code{\link[stats]{ts}}.
}
\details{
Les fonctions \code{ymd_ts()} et \code{dmy_ts()} permettent de convertir des tableaux (par exemple
des \code{\link[base]{data.frame}}) en \code{\link[stats]{ts}}. Pour cela il faut dans le tableau d'entrée
une colonne contenant les dates qui est au format YYYY-MM-JJ (fonction \code{ymd_ts()}) ou au format JJ-MM-YYYY
(fonction \code{dmy_ts()}). Le séparateur utilisé entre le jour, le mois et l'année peut être modifié par le
paramètre \code{sep_date}.
}
\examples{
x <- data.frame(dates=paste(rep(2000:2002,each=4),rep(c("01","04","07","10"),2),"01",sep="-"),donnees=1:12,stringsAsFactors = FALSE)
ymd_ts(x)
}
