% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facteurstatique.R
\encoding{UTF-8}
\name{facteurStatique}
\alias{facteurStatique}
\title{Analyse factorielle statique}
\usage{
facteurStatique(
  data,
  date_deb,
  date_fin,
  normalise = FALSE,
  retard_contrib = -1
)
}
\arguments{
\item{data}{un objet de type \code{\link[stats]{ts}} contenant les données.}

\item{date_deb}{date de début d'estimation.}

\item{date_fin}{date de fin d'estimation.}

\item{normalise}{booléen indiquant s'il faut renormaliser les données à une moyenne 100 et écart-type 10.}

\item{retard_contrib}{nombre de retards à prendre en compte pour le calcul des contribs.}
}
\value{
Un objet de même type que celui en entrée.
}
\description{
Fonction qui permet de calculer un indicateur synthétique par une analyse factorielle statique.
}
\examples{
series_cvs <- c(`Capacité d'épargne actuelle` = "000857195",
                `Capacité d'épargne future` = "000857198",
                `Évolution future du chômage` = "000857190",
                 `Niveau de vie futur en France` = "000857189",
                `Niveau de vie passé en France` = "000857188",
                `Opportunité de faire des achats importants`= "000857193",
                `Situation financière future` = "000857197",
                 `Situation financière passée` = "000857196")
data <- lectureBDM(series_cvs)
facteurStatique(data)
}
