% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_interactifs.R
\encoding{UTF-8}
\name{hc_ts}
\alias{hc_ts}
\alias{hc_lines}
\alias{hc_stocks}
\title{Graphiques interactifs d'une série temporelle avec highcharter}
\usage{
hc_lines(
  data,
  titre = NULL,
  sous_titre = NULL,
  legende = NULL,
  affiche_legende = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  outDec = ",",
  useHTML = FALSE,
  digits = 1
)

hc_stocks(
  data,
  titre = NULL,
  sous_titre = NULL,
  legende = NULL,
  affiche_legende = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  outDec = ",",
  useHTML = FALSE,
  type = NULL,
  color = NULL,
  digits = 1
)
}
\arguments{
\item{data}{un objet de type \code{\link[stats]{ts}} contenant une ou plusieurs séries à grapher.}

\item{titre}{titre du graphique (par défaut pas de titre).}

\item{sous_titre}{sous-titre du graphique (par défaut pas de sous-titre).}

\item{legende}{légende du graphique. Par défaut le nom des variables dans l'objet \code{data} en entrée. S'il n'y a qu'une série
à tracer alors aucune légende n'est affichée.}

\item{affiche_legende}{booléen indiquant si l'on affiche la légende.}

\item{x_lab}{titre axe des abscisses (par défaut aucun titre).}

\item{y_lab}{titre axe des ordonnées (par défaut "Date").}

\item{outDec}{séparateur décimal utilisé pour dans la légende des axes (par défaut la virgule).}

\item{useHTML}{Booléen indiquant si l'on veut utiliser du code HTML dans les titres.}

\item{digits}{nombre de chiffres après la virgule.}

\item{type}{vecteur contenant le type à utiliser pour les séries (par défaut \code{"lines"}).}

\item{color}{vecteur contenant les couleurs à utiliser (par défaut on garde les couleurs de highchart).}
}
\value{
Un graphique \code{\link[highcharter]{highcharter}}.
}
\description{
Fonctions qui permettent de faire des graphiques interactifs de séries temporelles à partir d'un objet de type \code{\link[stats]{ts}}.
}
\examples{
data <- lectureBDM("001585942","001585980")
legende <- c("Carnets de commandes globaux","Carnets de commandes étrangers")
titre <- "Soldes d'opinion sur les carnets de commandes dans l'industrie manufacturière"
hc_lines(data, titre = titre, legende = legende)
hc_stocks(data, titre = titre, legende = legende)
}
