% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_import.R
\encoding{UTF-8}
\name{lectureBDM}
\alias{lectureBDM}
\title{Téléchargement d'une série de la BDM Insee}
\usage{
lectureBDM(idbank, ...)
}
\arguments{
\item{idbank}{Un vecteur d'un ou plusieurs idbank à télécharger.}

\item{...}{D'autres vecteurs d'un ou plusieurs idbank à télécharger.}
}
\value{
Retourne un objet de type \code{ts} ou \code{mts}, en fonction du nombre d'idbank en entrée,
de la même fréquence que les séries en entrée.
}
\description{
Fonction qui permet d'importer des séries de la Banque de données macro-économiques (BDM) de l'Insee à partir de son idbank.
L'idbank d'une série peut être trouvé sur le site de la bdm (\url{https://www.bdm.insee.fr/bdm2/index}).
permet également l'utilisation de régressions quantiles.
}
\details{
La fonction permet de télécharger des séries temporelles présentes la Banque de données macro-économiques (BDM) de l'Insee
à partir de son idbank. Elle utilise pour cela le service web SDMX de l'Insee (voir \url{http://www.bdm.insee.fr/bdm2/statique?page=sdmx}
pour plus d'informations). Elle utilise le package \pkg{rsdmx} pour la lecture des données et \pkg{reshape2}
pour la mise en forme. Les idbank en paramètre doivent correspondre à des séries qui ont la même périodicité.
Si un des idbank en entrée n'est pas trouvé alors la fonction renvoit également un message d'avertissement.
}
\examples{

# Pour télécharger l'IPI manufacturier CVS : 001654241
lectureBDM("001654241")
# Pour télécharger les IPI dans l'industrie manufacturière et dans l'agroalimentaire :
lectureBDM(c("001654241","001654289"))
# ou bien :
lectureBDM("001654241","001654289")

}
