% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_import.R
\name{ts2xls}
\alias{ts2xls}
\title{Exporter données ts vers un fichier excel}
\usage{
ts2xls(x, file, sheet = "Feuille 1", format = "dd/mm/yyyy")
}
\arguments{
\item{x}{un objet de type \code{\link[stats]{ts}}.}

\item{file}{nom du fichier excel (avec extension .xls ou .xlsx)}

\item{sheet}{nom de la feuille}

\item{format}{format en sortie dans Excel}
}
\description{
Fonction qui permet d'exporter des données \code{ts} vers un fichier excel
en créant une colonne date qui sera au format date.
}
