% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\encoding{UTF-8}
\name{ts_utils}
\alias{ts_utils}
\alias{ev}
\alias{ga}
\title{Évolution et glissement annuel}
\usage{
ev(x)

ga(x)
}
\arguments{
\item{x}{un objet de type \code{\link[stats]{ts}} ou \code{\link[xts]{xts}}.}
}
\value{
Un objet de même type que celui en entrée.
}
\description{
Fonctions qui permettent, pour une série temporelle, de calculer l'évolution par rapport à la
dernière période (\emph{i.e.} : évolution mensuelle pour des séries mensuelles, évolution trimestrielle
 pour des séries trimestrielles, etc. ; avec \code{ev}) ou le glissement annuel
(avec \code{ga()}) .
}
\examples{
x  <-  ts(1:10, frequency = 4, start = c(1959, 2))
ev(x)
ga(x)
}
