% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\encoding{UTF-8}
\name{volatilite}
\alias{volatilite}
\title{Volatilité}
\usage{
volatilite(x, lag = 1)
}
\arguments{
\item{x}{un objet de type \code{\link[stats]{ts}} ou \code{\link[xts]{xts}}.}

\item{lag}{nombre de retards à utiliser pour le calcul de la différence (par défaut \code{lag = 1} : on calcule une
différence mensuelle pour les séries mensuelles).}
}
\value{
Un objet de même type que celui en entrée.
}
\description{
Fonction qui permet de calculer la volatité d'une série temporelle (i.e. : l'écart-type de sa différence). Les
valeurs manquantes ne sont pas prises en compte.
}
\examples{
x  <-  ts(1:10, frequency = 4, start = c(1959, 2))
volatilite(x)
}
