% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_import.R
\encoding{UTF-8}
\name{ctrl_v}
\alias{ctrl_v}
\title{Exporter des données vers le presse-papier}
\usage{
ctrl_v(data)
}
\arguments{
\item{data}{les données à exporter.}
}
\description{
Fonction qui permet d'exporter des données (un vecteur ou un \code{\link{data.frame}} par exemple)
vers le presse-papier (pour ensuite les coller dans un tableur par exemple).
}
\details{
Il est possible qu'il y ait une ligne en trop si la fonction est utilisée sous Mac Os.
}
\examples{
ctrl_v(c("line 1", "line 2 \\n and line 3"))
}
\seealso{
\code{\link{ctrl_c}} pour importer depuis le presse-papier.
}
